[#33] Refactored routes into separate file
