+++ cmd/server/server.go
@@ -1,8 +1,8 @@
 package server
 
 import (
-	"encoding/json"
 	"github.com/gorilla/mux"
+	"log"
 	"net/http"
 )
 
@@ -12,28 +12,26 @@ type api struct {
 
 type Server interface {
 	Router() http.Handler
+	SetRouter(*mux.Router)
 }
 
-func NewServer() Server {
+func NewServer() (Server, error) {
 	api := &api{}
 
-	router := mux.NewRouter()
-	api.router = router
+	error := mapRoutes(api)
 
-	router.HandleFunc("/rules", api.getRules).Methods(http.MethodGet)
+	if error != nil {
+		log.Printf("failed, lul %v", error)
+		return nil, error
+	}
 
-	return api
+	return api, nil
 }
 
 func (api *api) Router() http.Handler {
 	return api.router
 }
 
-func (api *api) getRules(writer http.ResponseWriter, request *http.Request) {
-	gophers := map[string]string{
-		"response": "nope, chuck testa",
-	}
-
-	writer.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(writer).Encode(gophers)
+func (api *api) SetRouter(r *mux.Router) {
+	api.router = r
 }
